Return data frame with R, Bioc versions and supported platforms

Return data frame with R, Bioc versions and supported platforms

Usage

available_df()

Value

A data frame

Examples

available_dates()
#>  [1] "2019-03-14" "2019-05-05" "2019-07-22" "2019-12-19" "2020-03-12"
#>  [6] "2020-04-27" "2020-06-22" "2020-08-20" "2020-10-30" "2021-02-26"
#> [11] "2021-04-01" "2021-05-29" "2021-08-03" "2021-10-28" "2022-01-16"
#> [16] "2022-04-19" "2022-06-22" "2022-08-22" "2022-10-20" "2022-12-20"
#> [21] "2023-02-13" "2023-04-01" "2023-06-15" "2023-08-15" "2023-10-30"
#> [26] "2023-12-30" "2024-02-29" "2024-04-29" "2024-06-14" "2024-08-19"
#> [31] "2024-10-01" "2024-12-14" "2025-01-14" "2025-01-24" "2025-01-27"
#> [36] "2025-02-03" "2025-02-10"