Create a Startup Script to Launch an Editor Inside a Nix Shell

Create a Startup Script to Launch an Editor Inside a Nix Shell

Usage

make_launcher(editor, project_path)

Arguments

editor

Character, the command to launch the editor. See the "Details" section below for more information.

project_path

Character, where to write the launcher, for example "/home/path/to/project". The file will thus be written to the file "/home/path/to/project/start-editor.sh". If the folder does not exist, it will be created.

Value

Nothing, writes a script.

Details

This function will write a launcher to start an IDE inside of a Nix shell. With a launcher, you only need to execute it instead of first having to drop into the shell using nix-shell and then type the command to run the IDE. You don’t need this launcher if you use direnv; see vignette("configuring-ide") for more details.

Examples

available_dates()
#>   [1] "2019-03-14" "2019-05-05" "2019-07-22" "2019-12-19" "2020-03-12"
#>   [6] "2020-04-27" "2020-08-20" "2020-10-30" "2021-02-26" "2021-04-01"
#>  [11] "2021-05-29" "2021-08-03" "2021-10-28" "2022-01-16" "2022-04-19"
#>  [16] "2022-06-22" "2022-08-22" "2022-10-20" "2022-12-20" "2023-02-13"
#>  [21] "2023-04-01" "2023-06-15" "2023-08-15" "2023-10-30" "2023-12-30"
#>  [26] "2024-02-29" "2024-04-29" "2024-06-14" "2024-08-19" "2024-10-01"
#>  [31] "2024-12-14" "2025-01-14" "2025-01-24" "2025-01-27" "2025-02-03"
#>  [36] "2025-02-10" "2025-02-17" "2025-02-24" "2025-02-28" "2025-03-03"
#>  [41] "2025-03-10" "2025-03-17" "2025-03-24" "2025-03-31" "2025-04-07"
#>  [46] "2025-04-11" "2025-04-14" "2025-04-16" "2025-04-29" "2025-05-05"
#>  [51] "2025-05-16" "2025-05-19" "2025-05-26" "2025-06-02" "2025-06-09"
#>  [56] "2025-06-13" "2025-06-24" "2025-07-02" "2025-07-07" "2025-07-14"
#>  [61] "2025-07-21" "2025-07-28" "2025-08-04" "2025-08-11" "2025-08-18"
#>  [66] "2025-08-25" "2025-09-01" "2025-09-04" "2025-09-09" "2025-09-11"
#>  [71] "2025-09-16" "2025-09-22" "2025-09-29" "2025-10-07" "2025-10-14"
#>  [76] "2025-10-20" "2025-10-27" "2025-11-01" "2025-11-03" "2025-11-10"
#>  [81] "2025-11-24" "2025-12-02" "2025-12-08" "2025-12-15" "2025-12-22"
#>  [86] "2025-12-29" "2026-01-05" "2026-01-14" "2026-01-19" "2026-01-26"
#>  [91] "2026-02-02" "2026-02-16" "2026-02-23" "2026-03-02" "2026-03-09"
#>  [96] "2026-03-11" "2026-03-16" "2026-03-23" "2026-03-30" "2026-04-06"
#> [101] "2026-04-13" "2026-04-22" "2026-04-30" "2026-05-04" "2026-05-11"
#> [106] "2026-05-18" "2026-05-25" "2026-06-09" "2026-06-15" "2026-06-22"
#> [111] "2026-07-27" "2026-08-03" "2026-08-10" "2026-08-24" "2026-08-31"